new tests passing
