+++ src/fem/block/element/composite/surface/triangle/test.rs
@@ -28,15 +28,23 @@ fn get_reference_coordinates() -> ReferenceNodalCoordinates<N>
         // [0.0, 0.0, 0.0],
         // [1.0, 0.0, 0.0],
         // [0.0, 1.0, 0.0],
-        // [0.0, 0.5, 0.0],
+        // OOPS [0.0, 0.5, 0.0],
         // [0.5, 0.5, 0.0],
         // [0.0, 0.5, 0.0]
+
+        // [1.0, 0.0, 0.0],
+        // [0.0, 1.0, 0.0],
+        // [0.0, 0.0, 0.0],
+        // [0.5, 0.5, 0.0],
+        // [0.0, 0.5, 0.0],
+        // [0.5, 0.0, 0.0]
+
+        [0.0, 0.0, 0.0],
         [1.0, 0.0, 0.0],
         [0.0, 1.0, 0.0],
-        [0.0, 0.0, 0.0],
+        [0.5, 0.0, 0.0],
         [0.5, 0.5, 0.0],
-        [0.0, 0.5, 0.0],
-        [0.5, 0.0, 0.0]
+        [0.0, 0.5, 0.0]
     ])
 }
